Reference

Reference

Mimi.@defcomp — Macro.

@defcomp name begin

Define a new component.

source

Mimi.setindex — Function.

setindex(m::Model, name::Symbol, count::Int)

Set the values of Model's' index name to integers 1 through count.

source

setindex{T}(m::Model, name::Symbol, values::Vector{T})

Set the values of Model's index name to values.

source

setindex{T}(m::Model, name::Symbol, valuerange::Range{T})

Set the values of Model's index name to the values in the given range valuerange.

source

Mimi.addcomponent — Function.

addcomponent(m::Model, t, name::Symbol=t.name.name; before=nothing,after=nothing)

Add a component of type t to a model.

source

Base.delete! — Function.

delete!(m::Model, component::Symbol

Delete a component from a model, by name.

source

Mimi.setparameter — Function.

setparameter(m::Model, component::Symbol, name::Symbol, value, dims)

Set the parameter of a component in a model to a given value. Value can by a scalar, an array, or a NamedAray. Optional argument 'dims' is a list of the dimension names of the provided data, and will be used to check that they match the model's index labels.

source

Set a component parameter as setparameter(reference, name, value).

source

Mimi.connectparameter — Function.

connectparameter(m::Model, component::Symbol, name::Symbol, parametername::Symbol)

Connect a parameter in a component to an external parameter.

source

connectparameter(m::Model, target_component::Symbol, target_name::Symbol, source_component::Symbol, source_name::Symbol; ignoreunits::Bool=false)

Bind the parameter of one component to a variable in another component.

source

connectparameter(m::Model, target::Pair{Symbol, Symbol}, source::Pair{Symbol, Symbol}; ignoreunits::Bool=false)

Bind the parameter of one component to a variable in another component.

source

Connect two components as connectparameter(reference1, name1, reference2, name2).

source

Connect two components as connectparameter(reference1, reference2, name).

source

Mimi.get_unconnected_parameters — Function.

get_unconnected_parameters(m::Model)

Return a list of tuples (componentname, parametername) of parameters that have not been connected to a value in the model.

source

Mimi.setleftoverparameters — Function.

setleftoverparameters(m::Model, parameters::Dict{Any,Any})

Set all the parameters in a model that don't have a value and are not connected to some other component to a value from a dictionary. This method assumes the dictionary keys are strings that match the names of unset parameters in the model.

source

Base.run — Function.

run(m::Model)

Run model m once.

source

Mimi.components — Function.

components(m::Model)

List all the components in model m.

source

Mimi.variables — Function.

variables(m::Model, componentname::Symbol)

List all the variables of componentname in model m.

source

variables(mi::ModelInstance, componentname::Symbol)

List all the variables of componentname in the ModelInstance 'mi'. NOTE: this variables function does NOT take in Nullable instances

source

Mimi.getdataframe — Function.

getdataframe(m::Model, componentname::Symbol, name::Symbol)

Return the values for variable name in componentname of model m as a DataFrame.

source

getdataframe(m::Model, comp_name_pairs::Pair(componentname::Symbol => name::Symbol)...)
getdataframe(m::Model, comp_name_pairs::Pair(componentname::Symbol => (name::Symbol, name::Symbol...)...)

Return the values for each variable name in each corresponding componentname of model m as a DataFrame.

source

Mimi.getindexcount — Function.

getindexcount(m::Model, i::Symbol)

Returns the size of index i in model m.

source

Mimi.getindexvalues — Function.

getindexvalues(m::Model, i::Symbol)

Return the values of index i in model m.

source

Mimi.getindexlabels — Function.

getindexlabels(m::Model, component::Symbol, x::Symbol)

Return the index labels of the variable or parameter in the given component.

source