Reference

Mimi.@defcomp — Macro.

@defcomp name begin

Define a new component.

source

Mimi.addcomponent — Function.

addcomponent(m::Model, t, name::Symbol=t.name.name; before=nothing,after=nothing)

Add a component of type t to a model.

source

Base.delete! — Function.

delete!(m::Model, component::Symbol)

Delete a component from a model, by name.

source

Mimi.addparameter — Function.

addparameter(m::Model, name::Symbol, value::NamedArray)

Add an array type parameter to the model, perferm dimension checking on the given NamedArray.

source

addparameter(m::Model, name::Symbol, value::AbstractArray)

Add an array type parameter to the model.

source

addparameter(m::Model, name::Symbol, value::AbstractArray, dims::Vector{Symbol})

Takes as input a regular array and a vector of dimension symbol names. Performs dimension name checks. Adds array type parameter to the model.

source

addparameter(m::Model, name::Symbol, value::Any)

Add a scalar type parameter to the model.

source

Mimi.setparameter — Function.

setparameter(m::Model, component::Symbol, name::Symbol, value)

Set the parameter of a component in a model to a given value.

source

Set a component parameter as setparameter(reference, name, value).

source

Mimi.connectparameter — Function.

connectparameter(m::Model, component::Symbol, name::Symbol, parametername::Symbol)

Connect a parameter in a component to an external parameter.

source

connectparameter(m::Model, target_component::Symbol, target_name::Symbol, source_component::Symbol, source_name::Symbol; ignoreunits::Bool=false)

Bind the parameter of one component to a variable in another component.

source

Connect two components as connectparameter(reference1, name1, reference2, name2).

source

Connect two components as connectparameter(reference1, reference2, name).

source

Mimi.get_unconnected_parameters — Function.

get_unconnected_parameters(m::Model)

Return a list of tuples (componentname, parametername) of parameters that have not been connected to a value in the model.

source

Mimi.setleftoverparameters — Function.

setleftoverparameters(m::Model, parameters::Dict{Any,Any})

Set all the parameters in a model that don't have a value and are not connected to some other component to a value from a dictionary.

source

Base.run — Function.

run(m::Model)

Run model m once.

source

Mimi.components — Function.

components(m::Model)

List all the components in model m.

source

Mimi.variables — Function.

variables(m::Model, componentname::Symbol)

List all the variables of componentname in model m.

source

variables(mi::ModelInstance, componentname::Symbol)

List all the variables of componentname in the ModelInstance 'mi'. NOTE: this variables function does NOT take in Nullable instances

source

Mimi.getdataframe — Function.

getdataframe(m::Model, componentname::Symbol, name::Symbol)

Return the values for variable name in componentname of model m as a DataFrame.

source

Mimi.getindexcount — Function.

getindexcount(m::Model, i::Symbol)

Returns the size of index i in model m.

source

Mimi.getindexvalues — Function.

getindexvalues(m::Model, i::Symbol)

Return the values of index i in model m.

source

Mimi.getindexlabels — Function.

getindexlabels(m::Model, component::Symbol, x::Symbol)

Return the index labels of the variable or parameter in the given component.

source