Reference

Mimi.components — Function.

components(m::Model)

List all the components in model m.

source

Mimi.variables — Function.

variables(m::Model, componentname::Symbol)

List all the variables of componentname in model m.

source

variables(mi::ModelInstance, componentname::Symbol)

List all the variables of componentname in the ModelInstance 'mi'. NOTE: this variables function does NOT take in Nullable instances

source

Mimi.addcomponent — Function.

Add a component to a model.

source

Mimi.setparameter — Function.

Set the parameter of a component in a model to a given value.

source

Set a component parameter as setparameter(reference, name, value).

source

Mimi.connectparameter — Function.

Bind the parameter of one component to a variable in another component.

source

Mimi.setleftoverparameters — Function.

Set all the parameters in a model that don't have a value and are not connected to some other component to a value from a dictionary.

source

Mimi.getdataframe — Function.

getdataframe(m::Model, componentname::Symbol, name::Symbol)

Return the values for variable name in componentname of model m as a DataFrame.

source

Base.run — Function.

run(m::Model)

Run model m once.

source

Mimi.@defcomp — Macro.

Define a new component.

source